Adjusted the design/layout of berlin.de/atlas/* AJAX-CityMap-Client. Extending the recently added /submit/*-Servlet.
FEATURES:
Added new MapBox based OSM Layer for self-styled map tiles.
Hovering of Layer-Select in AJAX-CityMap-Client.
The /submit-Servlet sends now notificatons after submission about the newly created dataset to either
- the mailbox of the logged in user/submitter if assigned (default)
- the given author-mailbox in the form (which is prioritized) 
- datasets (GeoObject-Topics) created via the /submit-Servlet now get an UUID-Random WebAlias assigned during creation

The notification mail should not send out the so called hiddenProps to the user. This is to be tested.

FIXES:
- fixed deep-linking to a search in AJAX-CityMap Clients
- revised the kiezatlas.setLayout()-Method
- fixed fahrinfo link configuration. link is now querying fahrinfo for addresses instead of bus and other stops
- fixed issue with authentication in /submit/*-Servlet

ADMINSTRATIVE CHANGES:
- /map/*-Servlet: removed the Google Maps API from both AJAX-CityMap Clients due to a suddenly invalid key
- added more descriptive comments into kiezatlas.js

IMPORTANT:
Requires DepaMehta2 in rev387 because the Form-Generator was extended to render hints above label/inputfield pair